requirements.txt
streamlit
pandas
pandas_datareader
plotly
python-dotenv